% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-helpers.R
\name{avom_ggplot2_defaults}
\alias{avom_ggplot2_defaults}
\title{Quickly set geoms defaults}
\usage{
avom_ggplot2_defaults(
  def_color = "#83a598",
  font_family = "Fira Sans",
  font_color = "#282828"
)
}
\arguments{
\item{def_color}{Default color for columns, points, lines, etc.}

\item{font_family}{Default font family for text.}

\item{font_color}{Default font color for text.}
}
\description{
Sets color and font family for common geoms.
}
\examples{
avom_ggplot2_defaults()
}
