% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes.R
\name{avom_palettes}
\alias{avom_palettes}
\title{Select a Color Palette}
\usage{
avom_palettes(palette = "gruvbox_light", n = NA)
}
\arguments{
\item{palette}{Name of a palette.}

\item{n}{Number of colors to extract. If \code{NA}, extracts all.}
}
\value{
A vector of color hexcodes.
}
\description{
Returns a vector of colors forming a cohesive palette. Current palettes include:
\itemize{
\item Nominal: \code{gruvbox_light}, \code{gruvbox_dark}
\item Diverging: \code{gruvbox_div}
}
}
\examples{
avom_palettes(palette = "gruvbox_light")
avom_palettes(palette = "gruvbox_dark", n = 5)
}
