% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-transforms.R
\name{count_pct}
\alias{count_pct}
\title{Count and Compute Relative Frequencies}
\usage{
count_pct(data, ..., wt = NULL, .by = NULL, labels = FALSE, suffix = "")
}
\arguments{
\item{data}{Dataset}

\item{...}{Additional arguments to be passed to \code{\link[dplyr:count]{dplyr::count()}}, including variables.}

\item{wt}{Optional weights}

\item{.by}{Optional variable to compute grouped percentages}

\item{labels}{If \code{TRUE}, returns values formatted for plotting`}

\item{suffix}{Set suffix for labels}
}
\value{
Dataframe with absolute and relative frequencies.
}
\description{
This is like \code{\link[dplyr:count]{dplyr::count()}} but also computes relative frequencies.
}
\details{
Uses \code{\link[dplyr:count]{dplyr::count()}} to compute absolute, possibly weighted, frequencies. Also computes relative frequencies and possibly labels for plots.
}
\examples{
count_pct(mtcars, cyl, labels = TRUE)
}
