% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes.R
\name{diverging_color_extractor}
\alias{diverging_color_extractor}
\title{Diverging Color Palette Extractor}
\usage{
diverging_color_extractor(palette, n)
}
\arguments{
\item{palette}{Name of a palette.}

\item{n}{Number of colors to extract.}
}
\value{
A vector of color hex codes
}
\description{
For diverging palettes, colors should be picked from the both ends, not sequentially.
}
\examples{
\dontrun{
   diverging_color_extractor(avom_palettes(palette = "gruvbox_div"), n = 3)
}
}
\keyword{internal}
