% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_project}
\alias{setup_project}
\title{Setup up Project Structure}
\usage{
setup_project()
}
\value{
Nothing
}
\description{
Sets up project structure.
}
\details{
For now, just creates folders I usually use.
}
\examples{
\dontrun{
setup_project()
}
}
