% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes.R
\name{avom_colors}
\alias{avom_colors}
\title{avom S7 palette class}
\usage{
avom_colors(name = character(0), type = character(0), colors = logical(0))
}
\arguments{
\item{name}{Name of the class, "avom_palette".}

\item{type}{Type of the of the palette: "nominal", "diverging" or "sequential".}

\item{colors}{Vector of colors.}
}
\value{
S7 object of class "avom_palette".
}
\description{
Creates a color palette, which is a S7 object of class "avom_palette", to be used in other functions.
}
\examples{
avom_colors(name = "gruvbox", type = "nominal", colors = c("#cc241d", "#98971a", "#458588", "#d79921", "#b16286", "#d65d0e", "#689d6a"))
}
