% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-helpers.R
\name{avom_ggsave}
\alias{avom_ggsave}
\title{Convenient wrapper around ggsave}
\usage{
avom_ggsave(
  filename,
  plot = ggplot2::last_plot(),
  width = 17,
  height = 8,
  scaling = 1,
  device = ragg::agg_png,
  ...
)
}
\arguments{
\item{filename}{Name of the exported file.}

\item{plot}{Exported plot}

\item{width}{Width of the plot. Defaults to 17 cm.}

\item{height}{Height of the plot. Defaults to 8 cm.}

\item{scaling}{Scaling factor for text. Defaults to 1.}

\item{device}{Graphical device to use. Defaults to \code{\link[ragg:agg_png]{ragg::agg_png()}}.}

\item{...}{Other arguments passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
}
\description{
A simple wrapper around \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}, which uses metric unit system and AGG graphical device by default.
}
\examples{
\dontrun{
mtcars_plot <- ggplot2::ggplot(mtcars,
                mapping = ggplot2::aes(x = mpg,
                                       y = hp)) +
ggplot2::geom_point()

avom_ggsave("mtcars-plot.png", mtcars_plot)
}
}
