% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes.R
\name{avom_palette}
\alias{avom_palette}
\title{Color Palettes for avom}
\usage{
avom_palette(palette = "catpuccin_latte")
}
\arguments{
\item{palette}{Name of the palette.}
}
\value{
S7 object of class "avom_palette".
}
\description{
Palettes for plotting.
}
\details{
Returns an S7 object representing color palette. The existing palettes are:
Nominal: catpuccin_latte, catpuccin_frappe, catpuccin_macchiato, catpuccing_mocha, gruvbox, gruvbox_light
Diverging: beach, royal_orange
Sequential: rising_reds, rising_blues
}
\examples{
avom_palette(palette = "catpuccin_latte")
}
