% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes.R
\name{palette_gen}
\alias{palette_gen}
\title{Color Palette Generator}
\usage{
palette_gen(palette = "catpuccin_latte", direction = 1)
}
\arguments{
\item{palette}{Name of the palette.}

\item{direction}{Order of colors. Set to -1 to reverse the order}
}
\value{
Function to be used in \code{\link[=scale_colour_avom]{scale_colour_avom()}} and \code{\link[=scale_fill_avom]{scale_fill_avom()}}.
}
\description{
Helper function to generate ggplot2 color palettes.
}
\examples{
palette_gen(palette = "catpuccin_latte", direction = 1)
}
