% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes.R
\name{scale_colour_avom}
\alias{scale_colour_avom}
\alias{scale_fill_avom}
\alias{scale_color_avom}
\title{Scale Color Palette for ggplot2}
\usage{
scale_colour_avom(palette = "catpuccin_latte", direction = 1, ...)

scale_fill_avom(palette = "catpuccin_latte", direction = 1, ...)

scale_color_avom(palette = "catpuccin_latte", direction = 1, ...)
}
\arguments{
\item{palette}{Name of the palette.}

\item{direction}{Order of colors. Set to -1 to reverse the order}

\item{...}{Other arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\value{
Function to be used together with \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
Set color palette for ggplot2 color and fill scales.
}
\examples{
ggplot2::ggplot(data = mtcars,
                mapping = ggplot2::aes(x = mpg, y = hp, color = as.factor(cyl))) +
         ggplot2::geom_point() +
         avom::scale_colour_avom("catpuccin_latte")
}
