% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes.R
\name{select_diverging}
\alias{select_diverging}
\title{Select Colors from Diverging Palettes}
\usage{
select_diverging(palette, n)
}
\arguments{
\item{palette}{Name of the palette.}

\item{n}{Number of colors to select.}
}
\value{
Vector of colors.
}
\description{
Helper function to select colors from diverging palettes in correct order.
}
\examples{
select_diverging(palette = c("#264653", "#2a9d8f", "#e9c46a", "#f4a261", "#e76f51"), n = 4)
}
