% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_project}
\alias{setup_project}
\title{Setup up Project Structure}
\usage{
setup_project(air = TRUE, pak = TRUE)
}
\arguments{
\item{air}{Sets up air.toml for Air formatter}

\item{pak}{Installs pak (sets it up to work with renv if present)}
}
\value{
Nothing
}
\description{
Sets up project structure.
}
\details{
Creates folders I usually use. Optionally also sets up air and pak.
}
\examples{
\dontrun{
setup_project()
}
}
