% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-themes.R
\name{theme_avom}
\alias{theme_avom}
\title{ggplot2 theme from avom}
\usage{
theme_avom(font_family = "Fira Sans", font_size = 10, ...)
}
\arguments{
\item{font_family}{Font family. Defaults to "Fira Sans".}

\item{font_size}{Baseline size for text. Defaults to 10.}

\item{...}{Other arguments passed to \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}.}
}
\description{
ggplot2 theme based on gruvbox color palette and Fira Sans font.
}
\examples{
\dontrun{
ggplot2::ggplot(data = mtcars,
                mapping = ggplot2::aes(x = mpg, y = hp)) +
         ggplot2::geom_point() +
         theme_avom()
         }
}
