% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-themes.R
\name{theme_avom}
\alias{theme_avom}
\title{ggplot2 theme from avom}
\usage{
theme_avom(
  panel.grid.major = ggplot2::element_line(),
  panel.grid.minor = ggplot2::element_blank(),
  ...
)
}
\arguments{
\item{panel.grid.major}{Controls major leading lines. Expects \code{\link[ggplot2:element]{ggplot2::element_line()}} line function.}

\item{panel.grid.minor}{Controls minor leading lines.}

\item{...}{Other arguments to be passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\description{
ggplot2 theme based on gruvbox color palette and Fira Sans font.
}
\examples{
ggplot2::ggplot(data = mtcars,
                mapping = ggplot2::aes(x = mpg, y = hp)) +
         ggplot2::geom_point() +
         theme_avom()
}
